+++ src/tests/resolvers/address-resolver-test.ts
@@ -1,0 +1,14 @@
+export const createAddress = `
+mutation($address: CreateAddressInput!) {
+  createAddress(address: $address) {
+    id,
+    cep,
+    city,
+    complement,
+    neighborhood,
+    state,
+    street,
+    streetNumber,
+    userId
+  }
+}`;
