+++ src/types/address-type.ts
@@ -1,0 +1,5 @@
+import { AddressModel } from '../model/address-model';
+
+export type GetAllAddressType = {
+  createAddress: AddressModel;
+};
